Fix Eclipse warnings
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/validator/trunk@1094739 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/validator/routines/UrlValidator.java
@@ -63,7 +63,7 @@ import java.util.regex.Pattern;
  *  </pre>
  *
  * @see
- * <a href='http://www.ietf.org/rfc/rfc2396.txt' >
+ * <a href="http://www.ietf.org/rfc/rfc2396.txt">
  *  Uniform Resource Identifiers (URI): Generic Syntax
  * </a>
  *
@@ -197,7 +197,7 @@ public class UrlValidator implements Serializable {
      *        ignore the contents of schemes.
      */
     public UrlValidator(String[] schemes) {
-        this(schemes, (long)0);
+        this(schemes, 0L);
     }
 
     /**
